+++ code/core/src/csf/index.ts
@@ -1,5 +1,3 @@
-import { toStartCaseStr } from './toStartCaseStr.ts';
-
 /**
  * Remove punctuation and illegal characters from a story ID.
  *
@@ -31,34 +29,7 @@ export const toId = (kind: string, name?: string) =>
 export const toTestId = (parentId: string, testName: string) =>
   `${parentId}:${sanitizeSafe(testName, 'test')}`;
 
-/** Transform a CSF named export into a readable story name */
-export const storyNameFromExport = (key: string) => toStartCaseStr(key);
-
-type StoryDescriptor = string[] | RegExp;
-export interface IncludeExcludeOptions {
-  includeStories?: StoryDescriptor;
-  excludeStories?: StoryDescriptor;
-}
-
-function matches(storyKey: string, arrayOrRegex: StoryDescriptor) {
-  if (Array.isArray(arrayOrRegex)) {
-    return arrayOrRegex.includes(storyKey);
-  }
-  return storyKey.match(arrayOrRegex);
-}
-
-/** Does a named export match CSF inclusion/exclusion options? */
-export function isExportStory(
-  key: string,
-  { includeStories, excludeStories }: IncludeExcludeOptions
-) {
-  return (
-    // https://babeljs.io/docs/en/babel-plugin-transform-modules-commonjs
-    key !== '__esModule' &&
-    (!includeStories || matches(key, includeStories)) &&
-    (!excludeStories || !matches(key, excludeStories))
-  );
-}
+export { isExportStory, storyNameFromExport, type IncludeExcludeOptions } from './export-story.ts';
 
 export interface SeparatorOptions {
   rootSeparator: string | RegExp;
