+++ code/lib/eslint-plugin/src/utils/csf-helpers.ts
@@ -1,2 +1,2 @@
-export type { IncludeExcludeOptions } from '../../../../core/src/csf/index.ts';
-export { isExportStory, storyNameFromExport } from '../../../../core/src/csf/index.ts';
+export type { IncludeExcludeOptions } from '../../../../core/src/csf/export-story.ts';
+export { isExportStory, storyNameFromExport } from '../../../../core/src/csf/export-story.ts';
